Add support for normalizing line endings (#44)
The previous change added a way to ignore line endings by forcing everything to be Unix style. This replaces that feature by letting the user choose whether they want Windows style, Unix style, or just leaving them alone. So if you want to include line endings in your diff, choose "Leave as-is" (which is the default.) If you choose this, when line endings are different, the diff viewer will show no red or yellow anywhere in the file. If you choose to normalize line endings, you'll never see diffs for files that have different line endings because the files are normalized before the comparison is done.

Note that this means that you can have, for example, all CRLF on the source side and all LF on the target side and if you choose either or Windows or Unix Style, you won't see any differences. The only time the new line endings will get persisted to the target side is if there are other changes in the file too. (This is the same behavior as all the other settings on the Formatting tab of the Settings dialog.)

So if you're looking to bulk convert the whole target to new line endings, you should either use another program (like notepad++) or you should wipe the target, do a comparison which will find that everything is different, and then update the target with all the changes.

Other changes:

- Version bump
- Removed the startup code that was wiping the local settings file with every run. This behavior is generally not desirable when debugging, but it can be easily added in when it is needed.